style: rename package
According to pep8, packages should use underscores
if it improves readability.

https://www.python.org/dev/peps/pep-0008/\#package-and-module-names

Closes #2
