+++ .coveragerc
@@ -1,4 +1,4 @@
-# This file is part of resourcealchemy.
+# This file is part of resource_alchemy.
 # https://github.com/TomNeyland/resource-alchemy
 
 # Licensed under the TBD license:
@@ -10,7 +10,7 @@ omit =
     *tests.py
 branch = True
 source =
-    resourcealchemy
+    resource_alchemy
 
 [report]
 exclude_lines =
