+++ .travis.yml
@@ -1,4 +1,4 @@
-# This file is part of resourcealchemy.
+# This file is part of resource_alchemy.
 # https://github.com/TomNeyland/resource-alchemy
 
 # Licensed under the TBD license:
