+++ tests/__init__.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
-# This file is part of resourcealchemy.
+# This file is part of resource_alchemy.
 # https://github.com/TomNeyland/resource-alchemy
 
 # Licensed under the TBD license:
@@ -9,29 +9,29 @@
 # Copyright (c) 2015, Tom Neyland <tcneyland+github@gmail.com>
 
 from preggy import expect
-import resourcealchemy
+import resource_alchemy
 from tests.base import TestCase
 
 
 class ExportsTestCase(TestCase):
 
-    def test_resourcealchemy_exports(self):
+    def test_resource_alchemy_exports(self):
 
         # Base Resource Classes
-        expect(resourcealchemy.ModelResource).not_to_be_null()
-        expect(resourcealchemy.ApiResource).not_to_be_null()
+        expect(resource_alchemy.ModelResource).not_to_be_null()
+        expect(resource_alchemy.ApiResource).not_to_be_null()
 
         # Basic fields
-        expect(resourcealchemy.Field).not_to_be_null()
-        expect(resourcealchemy.Relationship).not_to_be_null()
-        expect(resourcealchemy.ListRelationship).not_to_be_null()
-        expect(resourcealchemy.FilteredListRelationship).not_to_be_null()
-        expect(resourcealchemy.DateTimeField).not_to_be_null()
-        expect(resourcealchemy.ApiResource).not_to_be_null()
-        expect(resourcealchemy.IntervalField).not_to_be_null()
+        expect(resource_alchemy.Field).not_to_be_null()
+        expect(resource_alchemy.Relationship).not_to_be_null()
+        expect(resource_alchemy.ListRelationship).not_to_be_null()
+        expect(resource_alchemy.FilteredListRelationship).not_to_be_null()
+        expect(resource_alchemy.DateTimeField).not_to_be_null()
+        expect(resource_alchemy.ApiResource).not_to_be_null()
+        expect(resource_alchemy.IntervalField).not_to_be_null()
 
         # Basic Authorization Types
-        expect(resourcealchemy.FullAuthorization).not_to_be_null()
-        expect(resourcealchemy.ReadOnlyAuthorization).not_to_be_null()
-        expect(resourcealchemy.PropertyAuthorization).not_to_be_null()
-        expect(resourcealchemy.NoAuthorization).not_to_be_null()
+        expect(resource_alchemy.FullAuthorization).not_to_be_null()
+        expect(resource_alchemy.ReadOnlyAuthorization).not_to_be_null()
+        expect(resource_alchemy.PropertyAuthorization).not_to_be_null()
+        expect(resource_alchemy.NoAuthorization).not_to_be_null()
